+++ hawkey/src/package.h
@@ -40,7 +40,7 @@ const char *hy_package_get_packager(HyPackage pkg);
 const char *hy_package_get_reponame(HyPackage pkg);
 const char *hy_package_get_summary(HyPackage pkg);
 const char *hy_package_get_url(HyPackage pkg);
-unsigned long hy_package_get_epoch(HyPackage pkg);
+unsigned long long hy_package_get_epoch(HyPackage pkg);
 unsigned long long hy_package_get_medianr(HyPackage pkg);
 unsigned long long hy_package_get_rpmdbid(HyPackage pkg);
 unsigned long long hy_package_get_size(HyPackage pkg);
